Fix some bug

+++ framework/VertexArray.cpp
@@ -25,7 +25,7 @@ void VertexArray::AddBuffer(const VertexBuffer& vb, const VertexBufferLayout& la
     for(unsigned int i = 0; i < elements.size(); ++i) {
         const auto& element = elements[i];
         glEnableVertexAttribArray(0);
-        glVertexAttribPointer(i, element.count, element.type, element.normalized, layout.GetStride(), (const void*)offset);
+        glVertexAttribPointer(i, element.count, element.type, element.normalized, (size_t)layout.GetStride(), (const void*)offset);
         offset += element.count * VertexBufferElement::GetSizeOfType(element.type);
     }
 }
